Move `data/image.{hpp|cpp}` to `base`
