+++ src/assets/resource_loader.hpp
@@ -21,8 +21,8 @@
 #include "assets/duke_script_loader.hpp"
 #include "assets/palette.hpp"
 #include "base/array_view.hpp"
+#include "base/image.hpp"
 #include "data/audio_buffer.hpp"
-#include "data/image.hpp"
 #include "data/movie.hpp"
 #include "data/song.hpp"
 #include "data/sound_ids.hpp"
