+++ src/base/image.cpp
